Crate forest_genesis

Constants

EXPORT_SR_40

Functions

get_network_name_from_genesis

import_chain

Import a chain from a CAR file. If the snapshot boolean is set, it will not verify the chain state and instead accept the largest height as genesis.

initialize_genesis

read_genesis_header

Uses an optional file path or the default genesis to parse the genesis and determine if chain store has existing data for the given genesis.